feat: comments & paste/comment reactions; renames/refactors this is a LOT, i should've commited in stages but oh well, from what i can remember: - paste interactions (reactions) - comments - comment interactions (reactions) - rich text editor - move all report apis into report controller - rename size to sizeStored - rename trueSize to originalSize - fixed compression on paste compression     - implemented compression threshold - use longs instead of ints where possible in the entities - improve record view + remove redundant check for existing view

+++ SharpbinV3.Server/DTOs/Report/ReportQueryDto.cs
@@ -6,7 +6,7 @@ namespace SharpbinV3.Server.DTOs.Report
     {
         public Guid? ReporterUUID { get; init; }
         public Guid? UserUUID { get; init; }
-        public int? PastePID { get; init; }
+        public long? PastePID { get; init; }
         public string? PasteId { get; init; }
         public ReportTargetType? TargetType { get; init; }
         public ReportType? Type { get; init; }
